Version 8.2 introduces a new feature in the form of indirect supportive applications. These additional features are accessed through the main menu under the heading of Supp Apps. At present there is a single supplementary app called Filter which can be used to extract files from a folder if their notes or comments contain one of a range of user definable keywords.  This can typically be used to identify all .wav files in a folder which have not been analysed or in which it is was not possible to make a clear identification.  The same app can be used to re-merge re-analysed files into the main folder and to update the database accordingly. 8.2 will also progressively improve the user interface in situations where therre are multiple click type to access a range of features.  In Future right-clicking will bring up a context menu from which the appropriate action can be selected.
8.2.7320
Introduces the Filter supplementary App and also adds a right-click context menu when selecting one or more recordings in the List-By-Bats View.
If one or more recordings are selected in the List-ByBats View the context menu gives the choice of opening the Add Image dialog (the previous right-click function) or of exporting those .wav files to a user selected folder.  In this case the original files and the database are unaffected.